modified data type of interface

+++ src/main/java/com/robot/war/entity/interfaces/TypeOfAttack.java
@@ -1,6 +1,6 @@
 package com.robot.war.entity.interfaces;
 
 public interface TypeOfAttack {
-    public int lifeAttack();
-    public int shieldAttack();
+    public float lifeAttack();
+    public float shieldAttack();
 }
